feat: Car 객체 구현
public: randomPosition(), getName(), getPosition()
private: iForwardCondition(int randomNumber)
랜덤한 수를 받아서 조건에 맞으면 위치이동
